+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ dev
 * Added support for decompilation of unreachable functions.
 * Added support for disabling the emission of addresses in the generated code.
 * Added support for selecting functions to be decompiled.
+* Added support for selecting address ranges to be decompiled.
 
 0.3 (2016-07-17)
 ----------------